fix: catch exception in computed
When an exception is thrown during the execution of a computed signal,
the computation process of the signals state graph is interrupted. This
can disrupt the consistency of other signals.

Following Jotai's behavior
(see https://stackblitz.com/edit/vitejs-vite-i8u9jc24?file=src%2Fjotai.test.ts),
when an exception is thrown in a computed signal, record the exception in
the computed state and continue the state computation of other signals.
This exception will only be thrown when the computed signal is accessed `get`.